Re-organize modules and update `nestia.config.ts`
