+++ apps/nestjs-rest-api/src/main.ts
@@ -1,5 +1,5 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import { AppModule } from './modules/app/app.module';
 import { Logger } from '@nestjs/common';
 import * as swaggerUi from 'swagger-ui-express';
 import { readFileSync } from 'fs';
